Alignment of the original implementation

+++ mmdet3d/datasets/semantickitti_dataset.py
@@ -86,7 +86,7 @@ class SemanticKITTIDataset(Seg3DDataset):
         return self.data_list
 
     def get_seg_label_mapping(self, metainfo):
-        seg_label_mapping = np.zeros(metainfo['max_label'] + 1)
+        seg_label_mapping = np.zeros(metainfo['max_label'] + 1, dtype=np.int64)
         for idx in metainfo['seg_label_mapping']:
             seg_label_mapping[idx] = metainfo['seg_label_mapping'][idx]
         return seg_label_mapping
